requirements.txt
opencv-python
qrcode
requests
matplotlib
pillow
